app/models/recs.py
from pydantic import BaseModel
from typing import Optional, List
from .tmdb import TMDBMovieCard, TMDBMovieDetails

class TFIDFRecItem(BaseModel):
    title: str
    score: float
    tmdb: Optional[TMDBMovieCard] = None

class SearchBundleResponse(BaseModel):
    query: str
    movie_details: TMDBMovieDetails
    tfidf_recommendations: List[TFIDFRecItem]
    genre_recommendations: List[TMDBMovieCard]