requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1


numpy==2.1.0
scipy==1.13.1
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2

streamlit==1.36.0